let accumulators handle block sizes or # of blocks that exceed 32-bit int

+++ etomica.apps/etomica/virial/simulations/SimulationVirialOverlap.java
@@ -9,7 +9,6 @@ import java.io.IOException;
 import etomica.action.activity.ActivityIntegrate;
 import etomica.api.ISpecies;
 import etomica.data.DataPump;
-import etomica.exception.ConfigurationOverlapException;
 import etomica.integrator.IntegratorMC;
 import etomica.integrator.mcmove.MCMoveBoxStep;
 import etomica.integrator.mcmove.MCMoveManager;
@@ -173,7 +172,7 @@ public class SimulationVirialOverlap extends Simulation {
         }
     }
     
-    public void setAccumulatorBlockSize(int newBlockSize) {
+    public void setAccumulatorBlockSize(long newBlockSize) {
         blockSize = newBlockSize;
         for (int i=0; i<2; i++) {
             accumulators[i].setBlockSize(newBlockSize);
@@ -219,7 +218,7 @@ public class SimulationVirialOverlap extends Simulation {
                 integrators[i].getMoveManager().setEquilibrating(true);
             }
 
-            int oldBlockSize = blockSize;
+            long oldBlockSize = blockSize;
             // 1000 blocks
             long newBlockSize = initSteps*integratorOS.getNumSubSteps()/1000;
             if (newBlockSize < 1000) {
@@ -259,7 +258,7 @@ public class SimulationVirialOverlap extends Simulation {
         // run a short simulation to get reasonable MC Move step sizes and
         // (if needed) narrow in on a reference preference
         ai.setMaxSteps(initSteps);
-        int oldBlockSize = blockSize;
+        long oldBlockSize = blockSize;
         // 1000 blocks
         long newBlockSize = initSteps*integratorOS.getNumSubSteps()/1000;
         if (newBlockSize < 1000) {
@@ -326,7 +325,7 @@ public class SimulationVirialOverlap extends Simulation {
     public ActivityIntegrate ai;
     public IntegratorOverlap integratorOS;
     public double refPref;
-    protected int blockSize;
+    protected long blockSize;
 
   /*  public static void main(String[] args) {
